feat: Refactor SeleniumScraper to use ArrayList instead of List for memorials
The code changes in App.java update the type of the 'memorials' variable from List<Memorial> to ArrayList<Memorial>. This change is made in the scrapeMemorials method of the SeleniumScraper class. The purpose of this change is to specifically use an ArrayList implementation for the 'memorials' variable. This change is likely made to take advantage of the specific features and methods provided by the ArrayList class.

Recent user commits suggest that there were previous issues with dynamic loading and scraping of memorial links. The recent repository commits also indicate that there were updates made to the SeleniumScraper class to handle dynamic loading and scraping of memorial links.

Based on the recent commits and the code changes, it can be inferred that the developer is working on improving the functionality of the SeleniumScraper class to handle dynamic loading and scraping of memorial links. The change to use ArrayList instead of List for the 'memorials' variable is likely part of this improvement.

Note: The commit message has been modified to adhere to the established commit message conventions and to remove meta information.